Fill in HIR hash for associated opaque types
